Generic invert_batch
