+++ algebra/primefield/src/lib.rs
@@ -49,6 +49,7 @@ mod convert;
 pub mod fft;
 mod field_element;
 pub mod geometric_series;
+mod invert_batch;
 mod ops;
 mod prime_field;
 mod proth_field;
@@ -66,8 +67,7 @@ pub use zkp_u256::MontgomeryParameters;
 // The smallest 252-bit Proth field
 pub use proth_field::{FieldElement, Parameters};
 
-// TODO: Make part of Field
-pub use field_element::{invert_batch, invert_batch_src_dst};
+pub use invert_batch::{invert_batch, invert_batch_src_dst};
 
 // Re-exports dependencies that are part of the public interface
 pub use zkp_u256 as u256;
